+++ sickbeard/tv.py
@@ -1,4 +1,4 @@
-# Author: Nic Wolfe <nic@wolfeden.ca>
+﻿# Author: Nic Wolfe <nic@wolfeden.ca>
 # URL: http://code.google.com/p/sickbeard/
 #
 # This file is part of SickGear.
@@ -559,6 +559,8 @@ class TVShow(object):
         if self.dvdorder != 0:
             lINDEXER_API_PARMS['dvdorder'] = True
 
+        logger.log('%s: Loading all episodes from %s..' % (self.indexerid, sickbeard.indexerApi(self.indexer).name))
+
         try:
             t = sickbeard.indexerApi(self.indexer).indexer(**lINDEXER_API_PARMS)
             showObj = t[self.indexerid]
@@ -566,8 +568,6 @@ class TVShow(object):
             logger.log('%s timed out, unable to update episodes from %s' %
                        (sickbeard.indexerApi(self.indexer).name, sickbeard.indexerApi(self.indexer).name), logger.ERROR)
             return None
-
-        logger.log('%s: Loading all episodes from %s..' % (self.indexerid, sickbeard.indexerApi(self.indexer).name))
 
         scannedEps = {}
 
@@ -912,7 +912,7 @@ class TVShow(object):
         else:
             t = tvapi
 
-        myEp = t[self.indexerid]
+        myEp = t[self.indexerid, False]
         if None is myEp:
             logger.log('Show not found (maybe even removed?)', logger.WARNING)
             return False
@@ -931,7 +931,7 @@ class TVShow(object):
         self.imdbid = getattr(myEp, 'imdb_id', '')
 
         if getattr(myEp, 'airs_dayofweek', None) is not None and getattr(myEp, 'airs_time', None) is not None:
-            self.airs = myEp["airs_dayofweek"] + " " + myEp["airs_time"]
+            self.airs = ('%s %s' % (myEp['airs_dayofweek'], myEp['airs_time'])).strip()
 
         if getattr(myEp, 'firstaired', None) is not None:
             self.startyear = int(str(myEp["firstaired"]).split('-')[0])
@@ -1771,15 +1771,7 @@ class TVEpisode(object):
                 self.deleteEpisode()
             return
 
-        if not sickbeard.ALLOW_INCOMPLETE_SHOWDATA and None is getattr(myEp, 'episodename', None):
-            logger.log('This episode (%s - %sx%s) has no name on %s' %
-                       (self.show.name, season, episode, sickbeard.indexerApi(self.indexer).name))
-            # if I'm incomplete on TVDB but I once was complete then just delete myself from the DB for now
-            if -1 != self.indexerid:
-                self.deleteEpisode()
-            return False
-
-        if None is getattr(myEp, 'absolute_number', None):
+        if getattr(myEp, 'absolute_number', None) in (None, ''):
             logger.log('This episode (%s - %sx%s) has no absolute number on %s' %
                        (self.show.name, season, episode, sickbeard.indexerApi(self.indexer).name), logger.DEBUG)
         else:
@@ -1827,8 +1819,7 @@ class TVEpisode(object):
         self.indexerid = getattr(myEp, 'id', None)
         if None is self.indexerid:
             logger.log('Failed to retrieve ID from %s' % sickbeard.indexerApi(self.indexer).name, logger.ERROR)
-            if -1 != self.indexerid:
-                self.deleteEpisode()
+            self.deleteEpisode()
             return False
 
         # don't update show status if show dir is missing, unless it's missing on purpose
@@ -1845,30 +1836,30 @@ class TVEpisode(object):
         # if we don't have the file
         if not ek.ek(os.path.isfile, self.location):
 
-            today = datetime.date.today()
-            delta = datetime.timedelta(days=1)
-            show_time = network_timezones.parse_date_time(self.airdate.toordinal(), self.show.airs, self.show.network)
-            show_length = datetime.timedelta(minutes=helpers.tryInt(self.show.runtime, 60))
-            tz_now = datetime.datetime.now(network_timezones.sb_timezone)
-            future_airtime = (self.airdate > (today + delta) or
-                              (not self.airdate < (today - delta) and ((show_time + show_length) > tz_now)))
+            if self.status in [SKIPPED, UNAIRED, UNKNOWN, WANTED]:
+                today = datetime.date.today()
+                delta = datetime.timedelta(days=1)
+                show_time = network_timezones.parse_date_time(self.airdate.toordinal(), self.show.airs, self.show.network)
+                show_length = datetime.timedelta(minutes=helpers.tryInt(self.show.runtime, 60))
+                tz_now = datetime.datetime.now(network_timezones.sb_timezone)
+                future_airtime = (self.airdate > (today + delta) or
+                                  (not self.airdate < (today - delta) and ((show_time + show_length) > tz_now)))
 
-            # if this episode hasn't aired yet set the status to UNAIRED
-            if future_airtime and self.status in [SKIPPED, UNAIRED, UNKNOWN, WANTED]:
-                msg = 'Episode airs in the future, marking it %s'
-                self.status = UNAIRED
-
-            # if there's no airdate then set it to unaired (and respect ignored)
-            elif self.airdate == datetime.date.fromordinal(1):
-                if IGNORED == self.status:
-                    msg = 'Episode has no air date and marked %s, no change'
-                else:
-                    msg = 'Episode has no air date, marking it %s'
+                # if this episode hasn't aired yet set the status to UNAIRED
+                if future_airtime:
+                    msg = 'Episode airs in the future, marking it %s'
                     self.status = UNAIRED
 
-            # if the airdate is in the past
-            else:
-                if UNAIRED == self.status:
+                # if there's no airdate then set it to unaired (and respect ignored)
+                elif self.airdate == datetime.date.fromordinal(1):
+                    if IGNORED == self.status:
+                        msg = 'Episode has no air date and marked %s, no change'
+                    else:
+                        msg = 'Episode has no air date, marking it %s'
+                        self.status = UNAIRED
+
+                # if the airdate is in the past
+                elif UNAIRED == self.status:
                     msg = ('Episode status %s%s, with air date in the past, marking it ' % (
                         statusStrings[self.status], ','.join([(' is a special', '')[0 < self.season],
                                                               ('', ' is paused')[self.show.paused]])) + '%s')
@@ -1884,6 +1875,9 @@ class TVEpisode(object):
 
                 else:
                     msg = 'Not touching episode status %s, with air date in the past, because there is no file'
+
+            else:
+                msg = 'Not touching episode status %s, because there is no file'
 
             logger.log(msg % statusStrings[self.status], logger.DEBUG)
 
@@ -2060,42 +2054,26 @@ class TVEpisode(object):
             logger.log('%s: Not creating SQL queue - record is not dirty' % self.show.indexerid, logger.DEBUG)
             return
 
-        myDB = db.DBConnection()
-        rows = myDB.select(
-            'SELECT episode_id FROM tv_episodes WHERE showid = ? AND indexer=? AND season = ? AND episode = ?',
-            [self.show.indexerid, self.show.indexer, self.season, self.episode])
-
-        epID = None
-        if rows:
-            epID = int(rows[0]['episode_id'])
-
         self.dirty = False
-        if epID:
-            # use a custom update method to get the data into the DB for existing records.
-            return [
-                'UPDATE tv_episodes SET indexerid = ?, indexer = ?, name = ?, description = ?, subtitles = ?, '
-                'subtitles_searchcount = ?, subtitles_lastsearch = ?, airdate = ?, hasnfo = ?, hastbn = ?, status = ?, '
-                'location = ?, file_size = ?, release_name = ?, is_proper = ?, showid = ?, season = ?, episode = ?, '
-                'absolute_number = ?, version = ?, release_group = ? WHERE episode_id = ?',
-                [self.indexerid, self.indexer, self.name, self.description, ",".join([sub for sub in self.subtitles]),
-                 self.subtitles_searchcount, self.subtitles_lastsearch, self.airdate.toordinal(), self.hasnfo,
-                 self.hastbn,
-                 self.status, self.location, self.file_size, self.release_name, self.is_proper, self.show.indexerid,
-                 self.season, self.episode, self.absolute_number, self.version, self.release_group, epID]]
-        else:
-            # use a custom insert method to get the data into the DB.
-            return [
-                'INSERT OR IGNORE INTO tv_episodes (episode_id, indexerid, indexer, name, description, subtitles, '
-                'subtitles_searchcount, subtitles_lastsearch, airdate, hasnfo, hastbn, status, location, file_size, '
-                'release_name, is_proper, showid, season, episode, absolute_number, version, release_group) VALUES '
-                '((SELECT episode_id FROM tv_episodes WHERE showid = ? AND season = ? AND episode = ?)'
-                ',?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?);',
-                [self.show.indexerid, self.season, self.episode, self.indexerid, self.indexer, self.name,
-                 self.description,
-                 ",".join([sub for sub in self.subtitles]), self.subtitles_searchcount, self.subtitles_lastsearch,
-                 self.airdate.toordinal(), self.hasnfo, self.hastbn, self.status, self.location, self.file_size,
-                 self.release_name, self.is_proper, self.show.indexerid, self.season, self.episode,
-                 self.absolute_number, self.version, self.release_group]]
+        return [
+            'INSERT OR REPLACE INTO tv_episodes (episode_id, indexerid, indexer, name, description, subtitles, '
+            'subtitles_searchcount, subtitles_lastsearch, airdate, hasnfo, hastbn, status, location, file_size, '
+            'release_name, is_proper, showid, season, episode, absolute_number, version, release_group, '
+            'scene_absolute_number, scene_season, scene_episode) VALUES '
+            '((SELECT episode_id FROM tv_episodes WHERE indexer = ? AND showid = ? AND season = ? AND episode = ?)'
+            ',?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,'
+            '(SELECT scene_absolute_number FROM tv_episodes WHERE indexer = ? AND showid = ? AND season = ? AND episode = ?),'
+            '(SELECT scene_season FROM tv_episodes WHERE indexer = ? AND showid = ? AND season = ? AND episode = ?),'
+            '(SELECT scene_episode FROM tv_episodes WHERE indexer = ? AND showid = ? AND season = ? AND episode = ?));',
+            [self.show.indexer, self.show.indexerid, self.season, self.episode, self.indexerid, self.indexer, self.name,
+             self.description,
+             ",".join([sub for sub in self.subtitles]), self.subtitles_searchcount, self.subtitles_lastsearch,
+             self.airdate.toordinal(), self.hasnfo, self.hastbn, self.status, self.location, self.file_size,
+             self.release_name, self.is_proper, self.show.indexerid, self.season, self.episode,
+             self.absolute_number, self.version, self.release_group,
+             self.show.indexer, self.show.indexerid, self.season, self.episode,
+             self.show.indexer, self.show.indexerid, self.season, self.episode,
+             self.show.indexer, self.show.indexerid, self.season, self.episode]]
 
     def saveToDB(self, forceSave=False):
         """